Deprecates old CMFGen reader and moves new one
