+++ tardis/io/model/__init__.py
@@ -1,4 +1,4 @@
 from tardis.io.model.readers.stella import read_stella_model
 
 # from tardis.io.model.stella import read_stella_model
-from tardis.io.model.cmfgen import read_cmfgen_model
+from tardis.io.model.readers.cmfgen import read_cmfgen_model
