+++ tardis/io/model/readers/tests/test_cmfgen_reader.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 from pytest import fixture
 from astropy import units as u
-from tardis.io.model.cmfgen import read_cmfgen_model
+from tardis.io.model.readers.cmfgen import read_cmfgen_model
 
 MODEL_DATA_PATH = Path(__file__).parent / "data"
